nuked long obsolete ant-based build
A long, long time ago, before this project was moved to github,
the war was not building correctly from maven and we had this
ant script to compile and assemble the war.  Now that this
project is building correctly from maven we no longer need
this ant script.  This script is broken, anyway.
Updating the ant-based build to simulate what maven is now doing
would be costly with no real value.