corrected ParaBank spelling
The web app is spelled ParaBank but some recent additions started
misspelling this as Parabank.  The correct spelling can already
be found in various pages and messages.  Additionally,
the UI and documenation for SOAtest and Virtualize refer to this
as ParaBank.

This fixes all mispellings of ParaBank except for the jtest
scripts.