Switched JAX-RS JAXB provider from Jettison to Jackson to fix many issues
Summary:
The JSON and XML response messages are in agreement, sending the exact
same values as before.  However, some number and string values in the
JSON messages had the wrong type, contradicting the JSON Schema type
definitions in the Swagger and analogous XML Schema types in the WADL.
This problem has to do with an undesirable "feature" of Jettison where the JSON
type can switch between string or number depending on presence of a decimal.
This is one of many undesirable behaviors of Jettison that we've previously had to
work around but this particular behavior does not appear to be configurable.
In contrast, Jackson gives us all the expected behavior without much fuss.

Fixed properties incorrectly sent as JSON string instead of JSON number:
   Account.balance
   Transaction.amount
   Position.purchasePrice

Fixed properties incorrectly sent as JSON number instead of JSON string:
   Address.zipCode

The following properties were removed from the Swagger definition.  They were
representing return values of helper methods that were never actually being sent
nor intended to be sent in any responses (XML or JSON).  These helper methods
just returned a value that was a function of one or two other fields in the object.
This fix was needed because Jackson would be picking up on these properties
similar to the Swagger processor.  We removed other properties like this recently
but missed catching these other ones:
   Account.availableBalance
   Customer.fullName
   Customer.username
   Customer.password
   Position.costBasis

The WADL and XSDs for the XML responses have no changes.
The XML response messages have no changes.

Reviewers: benken

Reviewed By: benken

Differential Revision: http://phabricator.parasoft.com/D2743