Swagger - exclude the "host" field
This field is optional.  If absent, the host/ip{:port} is that
of the host serving the document which is preferred since it
makes the document portable.

The logic we were trying to use to guess the host and port is
flawed, tomcat specific, doesn't handle case where the container
is listening on multiple ports.  It also doesn't handle the case
where the request is proxied, like if the container is listening
on 8080 but requests to 80 are proxied/forwarded to 8080.